requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Data processing & import/export
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Images
pillow==10.1.0

# Config
python-decouple==3.8

# Optional: Prophet for advanced forecasting (install separately if needed)
# prophet==1.1.5

# Optional: ARIMA forecasting (lighter alternative)
statsmodels==0.14.1

# Reports & PDF
reportlab==4.0.9